+++ lib/read/cardano-wallet-read.cabal
@@ -53,6 +53,7 @@ library
     Cardano.Read.Ledger.Tx.CollateralInputs
     Cardano.Read.Ledger.Tx.CollateralOutputs
     Cardano.Read.Ledger.Tx.Eras
+    Cardano.Read.Ledger.Tx.ExtraSigs
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -74,7 +75,6 @@ library
     Cardano.Wallet.Read.Eras.KnownEras
     Cardano.Wallet.Read.Hash
     Cardano.Wallet.Read.Tx
-    Cardano.Wallet.Read.Tx.ExtraSigs
     Cardano.Wallet.Read.Tx.Fee
     Cardano.Wallet.Read.Tx.Gen
     Cardano.Wallet.Read.Tx.Gen.Address
